+++ noodles-bam/examples/bam_reheader.rs
@@ -24,7 +24,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     writer.write_header(&header)?;
     writer.write_reference_sequences(header.reference_sequences())?;
 
-    for result in reader.records() {
+    for result in reader.records(&header) {
         let record = result?;
         writer.write_record(&header, &record)?;
     }