+++ noodles-bam/examples/bam_split.rs
@@ -72,7 +72,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut writers = build_writers(header.read_groups())?;
     write_headers(&mut writers, &header)?;
 
-    for result in reader.records() {
+    for result in reader.records(&header) {
         let record = result?;
 
         if let Some(rg) = find_read_group(record.data())? {